+++ app/src/main/java/com/github/bumblebee202111/doubean/feature/groups/groupTab/GroupTabScreen.kt
@@ -192,14 +192,18 @@ private fun LazyListScope.tabActionsItem(
             group?.findTab(tabId)?.let { tab ->
                 //actions
                 Row {
-                    if (tab.isFavorite || tab.notificationPreferences != null) {
+                    // Only displayed when either condition meets:
+                    // 1. Favorited
+                    // 2. Record exists and enabled
+                    if (tab.isFavorite || tab.notificationPreferences?.notificationsEnabled == true) {
                         TabNotificationsButton(
                             notificationsEnabled = tab.notificationPreferences?.notificationsEnabled
                                 ?: false
                         ) {
                             onOpenAlertDialog()
                         }
-
+                    }
+                    if (tab.isFavorite) {
                         IconButton(removeFavorite) {
                             Icon(imageVector = Icons.Default.Star, contentDescription = null)
                         }