+++ app/src/main/java/com/github/bumblebee202111/doubean/feature/groups/groupTab/GroupTabViewModel.kt
@@ -70,7 +70,7 @@ class GroupTabViewModel @AssistedInject constructor(
         val tabId = tabId ?: return
         viewModelScope.launch {
             userGroupRepository.removeFavoriteTab(tabId)
-            shouldDisplayUnfavoritedTab = false
+            shouldDisplayUnfavoritedTab = true
         }
     }