+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Change Log
 
 ## Version 0.7.15(715)
 
+- Fix subtle issues of group/tab notifications button & tab favorite button
+
 ## Version 0.7.14(714)
 
 - Add pagination for RankList